// app/services/enums/contactInfoEndpoints.ts
export enum contactInfoEndpoints {
  CandidateProfileDetails = '/atsCandidateProfileDetails',
  CandidatePreDefinedTags = '/atsListPredefinedTags',
  ListAllCandidateTags = '/atsListAllCandidateTags',
  AddNewTag = '/atsAddNewTags',
  DeleteTag = '/atsDeleteCandidateTags',
  BlockContact = '/atsMarkAsSpam',
  EditParameters = '/atsEditParameters',
  DeleteParameter = '/atsDeleteParameters',
  AddNewNote = '/atsAddNewNotes',
  EditNote = '/atsEditNotes',
  DeleteNote = '/atsDeleteNotes',
}